moved spatial to its own package.
it was in util, but better to keep it in its own (first step to get rid
of util).
